DMBoundary: temporary fix to plex section boundary issue
DMPlexCreateSection isn't handling the case where
there are more dofs than components on a point properly.
There seems to be an ad hoc API that if you pass numcomps == 0
to DMAddBoundary, it constraints all components.  That's not good, but
I documented it and it avoided the plex problem until we can fix it.