PetscFn: an interface for nonlinear functions
PetscFn is an attempt to create a unified interface for
nonlinear functions as they appear in many existing PETSc interfaces.  A
PetscFn can be any nonlinear map from one space to another, in the same
way that a Mat is a linear map from one space to another
(`PetscFnApply(fn, x, y)`).

A scalar-valued function is a function, but it would be tedious
if every application of a scalar function had to unwrap
a vector (and potentially broadcast it).  PetscFn functions
that have a one-dimensional range can be used with a scalar subset of
the interface (`PetscFnScalarApply(fn, x, &z)`).

If PetscFn is to be useful, the amount of work
that it takes to implement a PetscFn should not be much greater
than the work it takes to implement callbacks right now in, e.g. SNES or
Tao.  If a PetscFn implementation (in this pull request there is only
PETSCFNSHELL, and operations are set with `PetscFnShellSetOperation()`)
only implements the scalar-function backend, the vector-valued functions
will still work.  (I am working on a computation graph implementation
of PetscFn chaining together other PetscFns, and things a greatly
simplified if I don't have to differentiate between scalar and vector
valued nodes in the graph).

```
    PetscFnCreate(comm, &fn);
    PetscFnSetSizes(fn, PETSC_DETERMINE, n, 1, M); /* scalar-valued function */
    PetscFnSetType(fn, PETSCFNSHELL);
    PetscFnShellSetOperation(fn, PETSCFNOP_SCALARAPPLY, myobjective);
    PetscFnShellSetOperation(fn, PETSCFNOP_SCALARDERIVATIVEVEC, mygradient);
    PetscFnCreateVecs(fn, NULL, &x, NULL, &y); /* create domain and range vectors */
                                               /* range vector is a single scalar */
    PetscFnScalarApply(fn, x, &z); /* works */
    PetscFnApply(fn, x, y); /* works */
    PetscFnCreateVecs(fn, NULL, &g, NULL, &y); /* gradient vector */
    PetscFnScalarGradient(fn, x, g); /* works */
    VecSet(y,1.);
    PetscFnJacobianMultAdjoint(fn, x, y, g); /* works for scalar and vector valued */
                                             /* functions, happens to be equal to */
                                             /* PetscFnScalarGradient if fn is */
                                             /* scalar valued */
```

A derivative of a function of some order is a tensor.  Rather than
introduce a tensor object, derivatives of PetscFn objects are available
in several froms, classified by how many variations are contracted
with the derivative tensor at a time:

- if all variations are contracted, the output is a scalar
- if all but one variation is contracted, the output is a vector
- if all but two variations are contracted, the output is a matrix

Because partial derivatives are useful in many settings, and because it
may be expensive to compute full derivatives and then restrict, partial
derivatives may by computed by passing in subsets with the variations.
For example, a PetscFn can be used for a classic PDE constrained
optimization constraint $A(u;m) = 0$ by specifying the domain of a to be
all $(u,m)$ vectors (say as VECNEST vectors).  We can ask for
$\partial_u A$ by passing in the index set of the state variables.

```
   IS subsets[2];

   subsets[0] = stateIS; /* we want the partial derivative wrt the state only */
   subsets[1] = NULL;    /* we want the full range of the Jacobian, NULL means all */

   PetscFnCreateVecs(A, NULL, &x, NULL, NULL); /* x is a (u,w) vector */
   PetscFnCreateVecs(A, stateIS, &uhat, NULL, &Juhat); /* variation on u and its range */
   PetscFnCreateVecs(A, stateIS, &Jadjv, NULL, &v); /* variation on u and its range */

   /* Jacobian action on uhat */
   PetscFnDerivativeVec(A,
                        x,
                        1, /* first derivative */
                        1, /* the derivative tensor is not transposed/swapped in any way, */
                           /* so none of the (in this case, one) variations is a range-space vector */
                        subsets,
                        &uhat, /* an array of variations, but in this case just one */
                        Juhat);

   subsets[0] = NULL;
   subsets[1] = stateIS;

   /* Jacobian adjoint action */
   PetscFnDerivativeVec(A,
                        x,
                        1, /* first derivative */
                        0, /* the derivative tensor *is* transposed: a range-space vector appears at */
                           /* position 0 in the variations */
                        subsets,
                        &v, /* an array of variations, but in this case just one */
                        Jadjv);
```

This is a flexible interface but with a learning curve: wrappers are
provided for common full derivatives, e.g.:

- `PetscFnJacobianMult(fn,x,xhat,Jxhat) -> PetscFnDerivativeVec(fn,x,1,1,NULL,&xhat,Jxhat)`
- `PetscFnJacobianMultAdjoint(fn,x,v,Jadjv) -> PetscFnDerivativeVec(fn,x,1,0,NULL,&v,Jadjv)`
- `PetscFnHessianMult(fn,x,a,b,Hab) -> PetscFnDerivativeVec(fn,x,2,2,NULL,[a,b],Hab)`
- `PetscFnHessianMultAdjoint(fn,x,v,a,Hadjva) -> PetscFnDerivativeVec(fn,x,2,0,NULL,[v,a],Hadjva)`
- `PetscFnScalarGradient(fn,x,g) -> PetscFnScalarDerivativeVec(fn,x,1,NULL,NULL,g)`

The derivative of a function is a function, so the derivative of a PetscFn can also be instantiated
as a PetscFn.  This would facilitate, e.g., passing the gradient of the
objective to SNES as an equation to be solved: `PetscFnScalarDerivativeFn(objfn,1,NULL,NULL,&gradfn)`

At this point I am very open to suggestions/improvements in this
feature.

- Some of the names are very long rights now (`PetscFnScalarDerivativeScalar()`),
- Local sub functions and/or projected subspace subfunctions seem like
  an obvious fit but I haven't added them to the interface yet
- Same with refined and coarsened functions: ultimately, I think a DM
  implementation of a PetscFn could do a lot of the things a DMSNES does
  now while making a clearer separation between the function and its
  solver
- This isn't a perfect fit for TS: functions would have to be indexed by
  time in a clear way, and when we build the Jacobian of an IFunction
  we are building two components of a full Jacobian and adding them
  together.