MatMatMult_SeqAIJ_SeqAIJ_Combined: refactor to be correct with realloc
The previous code assumed that PetscRealloc returns the same address,
but this is not possible in general.  It also assumed a fill of 8
instead of using the fill parameter passed in.  This version attains
equivalent performance in my tests with gcc-9 and clang-8, using more
const variables instead of running pointers.