MatMatMult_SeqAIJ_SeqAIJ_Combined: split into symbolic and numeric
This is a combined implementation, but conceivably one could call the
symbolic part before changing values in the input matrix, therefore
the combined part is more naturally attributable to the Numeric phase.
This commit also configures subsequent calls to use
MatMatMultNumeric_SeqAIJ_SeqAIJ_Sorted to avoid a memory leak and
because there is no need to recompute the symbolic part.