CMakePackage: execute by list of args instead of quoting + string concat
Quoting with string concatenation spoils use of special characters or
quotes in arguments.  It can be done reliably using shlex, but is
simpler and cleaner to pass the arguments via a list instead of via
shell escapes.  We use shlex.quote only when storing
pkg.conf.PACKAGENAME so that the can be pasted into a shell for
debugging purposes.

Commit-type: portability-fix