MatMPIAIJGetLocalMat: fix reuse case and clarify man page
This "optimization" was introduced in 70a9ba44dd1.
There's no need to create a new matrix with MatCreateSeqAIJWithArrays since it will share the underlying data structure
Better off to just take a reference on the local matrix and return it.
Extended the MAT_REUSE_CASE (MatCopy is a dummy operation if the two matrices are the same one): at least in this case we get an error if we are reusing the wrong matrix (sparsity pattern changed)