Ensure that copy to GPU in MatAssemblyEnd_SeqAIJCUSPARSE() actually happens when needed.
Before this commit, the CPU and GPU versions of the arrays might sometimes differ, but the MatSeqAIJCUSPARSECopyToGPU()
call would not actually do any copy, because it would see that A->offloadmask == PETSC_OFFLOAD_BOTH and assume that no
copy is needed. We would then end up with an assembled matrix in which the CPU and GPU copies of the matrix entries
differed, but would have A->offloadmask == PETSC_OFFLOAD_BOTH, indicating both versions are valid.