Extension to enable cmds to come from file.
Allow commands to come from a file instead of the standard mpicommands.
There is the `command:` keyword in the test language, but want avoid
chaining a bunch of bash commands into a single one which makes it hard
to debug and can give false positives from losing the return values
(like in original `filter:` implementation).

Used to test cmake build using pkg-config