First cut at hipification
This is a "dumb" hipification process following AMD guidelines.
It includes build system changes for HIP as well as the initial
conversion.  It is not meant to be elegant, but rather show the
verbosity that arises without having the appropriate abstractions in
place.

It now passes mpich-cxx-py3 and ci-linux-cuda-double.

Currently working on getting the mindless hipification that I did to
actually work.