Remove VecCUDASetPinnedMemoryMin() and document that this must be set via the options database.
The problem is that if VecCUDASetPinnedMemoryMin() is called before
VecSetType(), it tries to access VECCUDA-specific things that don't exist yet!
The type has to first be set to VECCUDA, but the host allocation will have
already happened once the type has been set, so the user calling
VecCUDASetPinnedMemoryMin() won't have the desired effect.
There is also another problem: If a user changes the threshold for a Vec that
has already been allocated on the host, the wrong free() implementation may be
used, as the choice of which one to use is made by examining the minimum pinned
memory size.