preparation for automatic generation of Fortran types
external marks functions for which an explicit interface is not
available. For these functions, the compiler cannot determine the
calling signature, hence the user need to take care of passing in the
correct types to avoid segmentation faults.

Without 'external', it should be possible to use these functions but the
compiler might complain. On the long term, it is advisable to have
interfaces defined as this allows the compiler to check for correct
arguments during compilation