interfaces where possible
some interfaces requires modern Fortran compiler (support for type(*)).
Fallback to external if this is not possible. Most of these functions
are typically not used from Fortran and are simply added as callback
according to src/snes/interface/ftn-custom/zsnesf.c, so the interfaces
are overkill. However, they will not hurt