Various updates for getting compilation to work
This is still not robust.  Currently cannot even get configure to pass
with ROCm 3.5.1.

Include changes to the sys level which compiled with ROCm 2.4

Things are not passing at the vec level because of error in the "C-mode"
include files at 2.4.  Upgraded to 3.5.1 but can't even basic compiles
to work.

Pushing becuase of interest and coordinating dev with
noether.colorado.edu.     This message will be squashed.