Miscellaneous bug fixes
Two bugs:

o GPUSERIALIZED=1 didn't actually do anything.  Fixed by adding an
argument to `config/gmakegentest.py` that actually configures this.
This should not affect the previous pipelines, but handles the
day-to-day usage.

o Forgot to include a newline character in comment so there was one test
that was not serialized.