Standardize error codes
If a job is killed for some reason, then bash will add 128 to the
error code resulting in confusion.  For example, if a job exits
with a cuda error code of 52, then the test harness will often
report and error code of 180 resulting in confusion.  This
standardizes the error code so that what the test harness reports
is what the actual error code is.