Run streams-like benchmark with -log_view to help users understand why their speedup is not as expected
Example output:
**************************************************************************************************************************************
The CPU streams rate 21418.1 (MB/s) on all (2) MPI ranks is not nearly two times the rate on half of the ranks 17270.5.
This indicates the memory bandwidth does not scale with the number of MPI ranks. You might be using more ranks than optimal.
See: https://petsc.org/release/faq/#what-kind-of-parallel-computers-or-clusters-are-needed-to-use-petsc-or-why-do-i-get-little-speedup
**************************************************************************************************************************************

Cannot run the benchmark after PetscLogView() because the logging has been deleted and one cannot create vectors at that point

Commit-type: feature
/spend 2h