PCGAMG: fix invalid access from redundent KSPSetFromOptions(smoother)
If the user writes -mg_levels_ksp_type richardson, for example, then the
following code would be referencing invalid cheb->members after
KSPSetFromOptions(smoother) changed the smoother from Chebyshev to
Richardson. But we don't need this call because KSPSetFromOptions is
called later in PCSetUp_MG, thus it's harmless for us to handle
Chebyshev here and allow for it to be changed later.