Add PCSetUseSymmetricForm() to help prevent bugs like PCASM with KSPCG
A flaw in this is if the user specifically set the PCASMType to restrict and then used KSPCG it would silently change what the user set.

To fix that we would need to track all PETSc parameters to see if they are default or set by user, yuck.

Commit-type: feature, usability
Thanks-to: Jed Brown
Thanks-to: Stefano Zampini