Config: do not add -use_gpu_aware_mpi 0 to test options
As there is no standard and reliable ways to tell whether an MPI installation is GPU-aware, we should just let it go and avoid adding -use_gpu_aware_mpi 0 to test harness.

On Sunspot, the old code thought the aurora-mpich was not gpu-aware. So when I ran test harness, it sliently added -use_gpu_aware_mpi 0 to test options.
Of course, I was unaware of that and thought it was doing gpu-aware testing, because aurora-mpich is indeed gpu-aware.

To avoid this gotcha, I think we should just remove the code and let users decide it themself.