Vec: Add VecDuplicatePreallocationCOO()
Typically you want preallocation on a residual vector,
which is a vector in the dual space.  In that sense
the COO preallocation data is like a very lightweight DM.
As long as temporary dual vectors are all created by
a chain of duplication from some original dual vector,
you can be sure that the COO data will be duplicated.
But the problem is that we have first order Richardson
-like methods that freely mix primal and dual vectors
as the same vector space.  Okay, fine, we can just put
preallocation COO data into our solution vectors, too.
But it also makes sense for algorithms to create temporary
vectors from a Jacobian using MatCreateVecs(), so now we
would have to override that method as well.  It is all
very brittle, so the only robust this is to
explicitly duplicate preallocation data.