Fix PETSCSFWINDOW for empty ranks
Using addresses as keys to windows has a problem when the buffers are NULL
because a process has no sf roots.

- We squeeze processes with no roots and no leaves out of the communicator used for windows
- We use both the root data address and the leaf data address to key the window, even though
  the window is only created on the root address, because this guarantees that communication
  events don't collide.  This can result in multiple windows being created over the same root data,
  but this is allowed in the MPI standard

Note that misidentfying PetscSF communication links is not a problem for
PETSCSFBASIC (using a full graph, I'm not sure about patterns like
allgatherv yet), because the only collision that can happen is on a process that has
no leaves an no roots, which does not participate in communication anyway,
so collision on the leaf does not present a problem.